The number of DB calls was to high especially in case of nested includes. New relic was showing ~100x db calls for includes. I had to fix it for our project and use our custom branch. But I have written few test cases and verified the tests. Please merge this if you think this looks good. 

Performance Improvement:
================
Tool: gobench (https://github.com/cmpxchg16/gobench)

Before Optimization:
================
Concurrent Clients which can be connected w/o error per single process (on my mac): ~300 but average response time was ~20 secs
Maximum Requests/sec per single process: ~47 req/s with 500 clients
Maximum Requests/sec with 3 cluster processes: ~78 req/s with 500 clients

After Optimization:
================
Note: Because of increase in processing, there's a negative impact
Concurrent Users which can be connected per single process (on my mac): ~200 but average response time was ~3 secs
Maximum Requests/sec per single process: ~77 req/s with 500 clients
Maximum Requests/sec with 3 cluster processes: ~147 req/s with 500 clients

Conclusion
================
Although there's a performance impact on nodejs/web layer, it is much easier to scale web application layer than Database. So, I think this is a good solution for now. Please merge & optimize it more if possible.