@raymondfeng 2015-05-13
@ningsuhen Great work! 

What's your take on the DB limit on the number of values for the `IN` cause?  For example, if the `through` table has many matching ids and it goes beyond the limit of `IN`, should we consider to do the queries in multiple calls and merge the results (a special sort of pagination) ? That will complicate the sorting through.