@raymondfeng 2015-05-13
I fixed a few issues and landed the PR. Thanks!

We'll continue to discuss the IN limit issue.