@ningsuhen 2015-05-14
Thanks @raymondfeng. 
Assuming IN array limit will be a problem majorly in HasManyThrough relations. Using loop for this is also a problem especially when we have to sort & limit results. Optimizing client side is ok only upto a certain limit. With large number of records, it won't work. 

I'm not sure of the implications but another way is to denormalize indexed data (of target model) onto the through model. That way, we can sort and limit the results in the through Model itself. This is of course assuming each through model has a valid target model object.